# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/produce": {
      "post": {
        "operationId": "ProduceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProduceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProduceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Produce"
        ]
      },
      "get": {
        "operationId": "ProduceController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ProduceStatus"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ProduceType"
            }
          },
          {
            "name": "farmId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProduceListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Produce"
        ]
      }
    },
    "/produce/{id}": {
      "get": {
        "operationId": "ProduceController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProduceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Produce"
        ]
      },
      "patch": {
        "operationId": "ProduceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProduceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProduceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Produce"
        ]
      },
      "delete": {
        "operationId": "ProduceController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProduceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Produce"
        ]
      }
    },
    "/orders": {
      "post": {
        "operationId": "OrderController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      },
      "get": {
        "operationId": "OrderController_findAll",
        "parameters": [
          {
            "name": "produceId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/OrderStatus"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "The ordered produce's type",
            "schema": {
              "$ref": "#/components/schemas/ProduceType"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/count": {
      "get": {
        "operationId": "OrderController_count",
        "parameters": [
          {
            "name": "produceId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/OrderStatus"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "The ordered produce's type",
            "schema": {
              "$ref": "#/components/schemas/ProduceType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderCountResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "OrderController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      },
      "delete": {
        "operationId": "OrderController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/{id}/confirm": {
      "patch": {
        "operationId": "OrderController_confirm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/orders/{id}/cancel": {
      "patch": {
        "operationId": "OrderController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    }
  },
  "info": {
    "title": "laicos API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "perPage": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 42
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "page",
          "perPage",
          "total",
          "totalPages"
        ]
      },
...
      "ProduceStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "PUBLISHED",
          "SOLD_OUT"
        ]
      },
      "ProduceType": {
        "type": "string",
        "enum": [
          "LOCAL",
          "EXPORT"
        ]
      },
      "CreateProduceDto": {
        "type": "object",
        "properties": {
          "farmId": {
            "type": "string",
            "format": "uuid",
            "description": "A farm the caller owns"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "example": "Yellow maize"
          },
          "actualQuantity": {
            "type": "number",
            "minimum": 0,
            "example": 500,
            "description": "Stock on hand, in `unit`. floatingQuantity starts equal to it and is managed by orders."
          },
          "unit": {
            "type": "string",
            "maxLength": 20,
            "example": "kg"
          },
          "pricePerUnit": {
            "type": "number",
            "minimum": 1,
            "example": 350.5,
            "description": "Price of one `unit`"
          },
          "imageUrl": {
            "type": "string",
            "format": "uri",
            "example": "https://cdn.example.com/maize.jpg"
          },
          "status": {
            "default": "DRAFT",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProduceStatus"
              }
            ]
          },
          "type": {
            "default": "LOCAL",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProduceType"
              }
            ]
          }
        },
        "required": [
          "farmId",
          "name",
          "actualQuantity",
          "unit",
          "pricePerUnit"
        ]
      },
      "ProduceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "farmId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Yellow maize"
          },
          "quantity": {
            "type": "number",
            "example": 500,
            "description": "Original stock listed, in `unit`. Immutable after creation."
          },
          "actualQuantity": {
            "type": "number",
            "example": 500,
            "description": "Stock not yet committed to a CONFIRMED order, in `unit`"
          },
          "floatingQuantity": {
            "type": "number",
            "example": 450,
            "description": "Stock still orderable: actual minus PENDING orders"
          },
          "unit": {
            "type": "string",
            "example": "kg"
          },
          "pricePerUnit": {
            "type": "string",
            "example": "350.50"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProduceStatus"
              }
            ]
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProduceType"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "farmId",
          "name",
          "quantity",
          "actualQuantity",
          "floatingQuantity",
          "unit",
          "pricePerUnit",
          "imageUrl",
          "status",
          "type",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProduceResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "OK"
          },
          "data": {
            "$ref": "#/components/schemas/ProduceDto"
          }
        },
        "required": [
          "message",
          "data"
        ]
      },
      "ProduceListResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "OK"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProduceDto"
            }
          },
          "metaData": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "message",
          "data",
          "metaData"
        ]
      },
      "UpdateProduceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "example": "Yellow maize"
          },
          "actualQuantity": {
            "type": "number",
            "minimum": 0,
            "example": 500,
            "description": "Stock on hand, in `unit`. floatingQuantity starts equal to it and is managed by orders."
          },
          "unit": {
            "type": "string",
            "maxLength": 20,
            "example": "kg"
          },
          "pricePerUnit": {
            "type": "number",
            "minimum": 1,
            "example": 350.5,
            "description": "Price of one `unit`"
          },
          "imageUrl": {
            "type": "string",
            "format": "uri",
            "example": "https://cdn.example.com/maize.jpg"
          },
          "status": {
            "default": "DRAFT",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProduceStatus"
              }
            ]
          },
          "type": {
            "default": "LOCAL",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProduceType"
              }
            ]
          }
        }
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "produceId": {
            "type": "string",
            "format": "uuid",
            "description": "A PUBLISHED produce"
          },
          "quantity": {
            "type": "number",
            "minimum": 1,
            "example": 50,
            "description": "In the produce's `unit`"
          }
        },
        "required": [
          "produceId",
          "quantity"
        ]
      },
      "OrderStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "CONFIRMED",
          "READY",
          "FULFILLED",
          "CANCELLED"
        ]
      },
      "OrderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "produceId": {
            "type": "string",
            "format": "uuid"
          },
          "farmId": {
            "type": "string",
            "format": "uuid"
          },
          "buyerId": {
            "type": "string",
            "format": "uuid",
            "description": "User.id of the buyer"
          },
          "quantity": {
            "type": "number",
            "example": 50,
            "description": "In the produce's `unit`"
          },
          "totalPrice": {
            "type": "string",
            "example": "17525.00"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderStatus"
              }
            ]
          },
          "cancellationReason": {
            "type": "string",
            "nullable": true,
            "description": "Set when the order is CANCELLED"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "produceId",
          "farmId",
          "buyerId",
          "quantity",
          "totalPrice",
          "status",
          "cancellationReason",
          "createdAt",
          "updatedAt"
        ]
      },
      "OrderResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "OK"
          },
          "data": {
            "$ref": "#/components/schemas/OrderDto"
          }
        },
        "required": [
          "message",
          "data"
        ]
      },
      "OrderListResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "OK"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderDto"
            }
          },
          "metaData": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "message",
          "data",
          "metaData"
        ]
      },
      "OrderStatusCountsDto": {
        "type": "object",
        "properties": {
          "PENDING": {
            "type": "number",
            "example": 3
          },
          "CONFIRMED": {
            "type": "number",
            "example": 2
          },
          "FULFILLED": {
            "type": "number",
            "example": 7
          },
          "CANCELLED": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "PENDING",
          "CONFIRMED",
          "FULFILLED",
          "CANCELLED"
        ]
      },
      "OrderCountDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 13
          },
          "byStatus": {
            "$ref": "#/components/schemas/OrderStatusCountsDto"
          }
        },
        "required": [
          "total",
          "byStatus"
        ]
      },
      "OrderCountResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "OK"
          },
          "data": {
            "$ref": "#/components/schemas/OrderCountDto"
          }
        },
        "required": [
          "message",
          "data"
        ]
      },
      "CancelOrderDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500,
            "example": "Harvest was damaged by rain"
          }
        },
        "required": [
          "reason"
        ]
      },
...
    }
  }
}